Getting rid of MultiChoiceCaller
